FEAT: add success page, minor tweaks
Added a Success page for the user to be routed to when they submit a Movie successfully.

Minor Tweaks:
Podcasts: Renamed to Podcast.

AddMovie Post: add implicit cast from MovieForm to Movie. Redirect user to Success action if insertion successful.

EditMovie, DeleteMovie Get: change the parameter name to match what is being passed from the view. Changed the parameter type to match changes to the Database and Generic Repo.

EditMovie, DeleteMovie Post: change the parameter type from MovieForm to Movie so that the MovieID, really the most important part of the model, it passed to the view. Change the actions upon successful update or delete to redirect to FilmList action to avoid duplicate code.