FEAT: basic crud views
Add basic CRUD Razor views. Minimal styling. All pages required for the assignment are in here. Will need styling updates as well as sanity checks.

Add partial views for the Edit and Delete buttons that appear on the FilmList view.

Add picture of Joel Hilton on the Index view.

Add a view that gives the user the link the Baconsale.

Update _ViewImports to bring in the ViewModels namepspace.